kgeo integration in  amat;fixed exit->pnewdt corrections in mapshape3D

+++ 0_NH_DEF/src/element/uvarm.f90
@@ -14,7 +14,7 @@
 
         character(len=80) :: CMNAME, ORNAME
         character(len=3) :: FLGRAY(15)
-        real :: UVAR(NUVARM), DIRECT(3,3), T(3,3), TIME(2)
+        DOUBLE PRECISION :: UVAR(NUVARM), DIRECT(3,3), T(3,3), TIME(2)
         integer :: ARRAY(15), JARRAY(15), JMAC(*), JMATYP(*), COORD(*)
         integer :: i1
